Create SyneXus.sol
SyneXus is a crypto tokenomics project structured to reward those who buy and hold tokens, to the detriment of those who sell.
SyneXus is totally community-centric, created with the aim of combining individual growth goals and expectations to aspire together to a higher goal, focused on safeguarding and conserving the planet. We are convinced that "connected" can improve our lives and at the same time, make the world a better place for everyone.